feat: switch to env bash
This shebang uses the env command to locate the bash executable in the
user's PATH. It is more portable because it allows the script to use the
bash shell that is found first in the PATH, which can vary between
different systems. This approach is particularly useful in environments
where the location of the bash executable might differ, such as on
different Linux distributions or when using virtual environments.